dataloaderfix

+++ dataloader.py
@@ -119,7 +119,7 @@ class CIFAR10SSL(datasets.CIFAR10):
             target = self.target_transform(target)
 
         target = torch.tensor(target)
-        return img, target
+        return img, target.long()
 
 
 class CIFAR100SSL(datasets.CIFAR100):
@@ -145,7 +145,7 @@ class CIFAR100SSL(datasets.CIFAR100):
             target = self.target_transform(target)
 
         target = torch.tensor(target)
-        return torch.tensorimg, target
+        return torch.tensorimg, target.long()
 
 class TransformFixMatch(object):
     def __init__(self, mean, std):
